+++ CHANGELOG.md
@@ -2,11 +2,30 @@
 
 All notable changes to this project will be documented in this file.
 
-## [2.14.4-pre-110.1](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.14.3...v2.14.4-pre-110.1) (2026-06-23)
+## [2.16.0](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.15.0...v2.16.0) (2026-06-22)
+
+### Added
+
+* adds in the custom block for the elevate fan ([7127f4e](https://github.com/SSmale/Duux-Home-Assistant/commit/7127f4eb403d5d6d917aee8098606d0b885f60a3))
+* adds the ionzer switch ([e890db2](https://github.com/SSmale/Duux-Home-Assistant/commit/e890db2b2bd1f76d97b0c07c00a36162c054b3e1))
+
+## [2.15.0](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.14.3...v2.15.0) (2026-06-20)
+
+### Added
+
+* adds tests for the project ([#109](https://github.com/SSmale/Duux-Home-Assistant/issues/109)) ([53d0968](https://github.com/SSmale/Duux-Home-Assistant/commit/53d0968070a1f061b9dab0756325a4fee5561584))
+
+## [2.15.0-pre-109.2](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.15.0-pre-109.1...v2.15.0-pre-109.2) (2026-06-20)
 
 ### Fixed
 
-* prevent UnboundLocalError in async_set_preset_mode ([#110](https://github.com/SSmale/Duux-Home-Assistant/issues/110)) ([acf6dd8](https://github.com/SSmale/Duux-Home-Assistant/commit/acf6dd874b695cd7c176a02f3d2a75ed81b780d4))
+* fixes code issues with select ([280bf29](https://github.com/SSmale/Duux-Home-Assistant/commit/280bf29f58a73ce205e85261f2680bbd658b9273))
+
+## [2.15.0-pre-109.1](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.14.3...v2.15.0-pre-109.1) (2026-06-20)
+
+### Added
+
+* adds tests for the project ([a35fff8](https://github.com/SSmale/Duux-Home-Assistant/commit/a35fff851cff4bf7f46a89394db20fefbe137e20))
 
 ## [2.14.3](https://github.com/SSmale/Duux-Home-Assistant/compare/v2.14.2...v2.14.3) (2026-06-20)
 
@@ -341,6 +360,3 @@ All notable changes to this project will be documented in this file.
 * [ ] Add sensor entities for power consumption
 * [ ] Local API support (if available)
 * [ ] Energy dashboard integration
-
-
-